pyproject.toml
[project]
name = "104"
version = "0.1.0"
description = ""
authors = [
    {name = "uncleben006",email = "uncleben006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "streamlit (>=1.42.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
